Functional GUI with manual control and displaying status
This should have been several commits. But in short, this adds:

- A functional QML interface for manual control. Each valve, pressure
controller and pump can be controlled.
- The QML ValveSwitch and PressureController objects have helper objects
to handle the signals & slots and interface to the backend
(ApplicationController). These classes are defined in C++, in the
guihelper.[h/cpp] files.
- The ApplicationController is aware of these helper objects: when they
are created in QML, they register themselves to the
ApplicationController so that it can keep track of which valve number
corresponds to what object, for example.
- The ApplicationController class is now a static singleton, so the
instance can be accessed from anywhere, C++ or QML. This makes it
possible / easy to interact with the backend from anywhere.