Working demo. Bidirectional communication works just like USB
Major issues with connection -- I am not sure whether this is due to Qt,
Linux or the ESP32, but service discovery does not work properly. The
current workaround is to look up the device's address and channel ID,
and hardcoding these in ApplicationController.

Along with fixing this, the next main task should be to reconcile
Communicator and BluetoothCommunicator. They could both be made to
inherit from a same base class, which would be inherited e.g. by
SerialCommunicator and BluetoothCommunicator. That way, only one or two
lines of code would need to be changed to switch between the two.