Dark mode!
Note: colors are manually specified in LogScreen.qml because there doesn't seem to be a direct way to map a Material color to a QColor. I.e. you cannot do "color: Material.Orange". At least, not that I've found. But something similar would be nice, perhaps via a helper function.