Manual control: auto-resize buttons
I finally figured out how to properly make use of QML Layouts (for the most part)
Buttons now will resize  when the window width is changed, up to a given size.

There is a warning on a binding loop on the Flickable's 'contentHeight', but I don't know how to fix that and it still works fine, so I left it for now.