Overhaul of multiplexer control logic
Both visual elements and control logic are moved to QML: while this goes against the idea that QML is reserved for GUI elements while the C++ backend handles complex functionality, it ensures that all aspects of the software that are specific to a given microfluidic chip are defined within QML, specifically in the GraphicalControl element.

The multiplexer valves and desired configurations (i.e. which outputs can be selected) are specified in a simple format, and the buttons are populated based on this.

The routineController now connects to this new QML/JS-based multiplexer controller, rather than to the applicationController.