Fix loop when pressure setpoint is changed
When the setpoint of a pressure controller was updated on the GUI, this triggered the same event that sends a new setpoint to the hardware. For some reason, it only sometimes resulted in infinite loops...
This should fix it, so that only the "moved" signal (emitted when the slider is updated by mouse, keyboard,...) results in an update to the hardware.